pizzaBot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
peewee==3.14.10
python-telegram-bot==13.11
pytz==2022.1
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.1